setup done for Droppable
